Got the readc3d working.
Had to work around Matlab's (not) reading of VAX D floating point
anymore... Apparently it used to, but doesn't any more...
So, I used a wrapper around the fread function in the file.

Had to add in a couple of utility files to be able to format the data
correclty.